rename folders to match site navigation
